Refactor types location and exclude from coverage
